+++ core/src/test/kotlin/sift/core/api/testdata/set1/ClassWithExtensionFunction.kt
@@ -3,6 +3,8 @@
 package sift.core.api.testdata.set1
 
 class ClassWithExtensionFunction {
+    val foo // should not show up under methods(declared) (default)
+        get() = MethodsWithTypes.Foo()
 
     fun List<Foo>.hello(foo: Foo): String = TODO()