+++ core/src/test/kotlin/sift/core/api/testdata/set1/MethodsWithTypes.kt
@@ -3,7 +3,6 @@
 package sift.core.api.testdata.set1
 
 class MethodsWithTypes {
-
     fun presentTypes(foo: Foo, bar: Bar) = Unit
     fun mixedTypes(foo: Foo, s: String) = Unit
     fun noPresentTypes(i: Int, s: String) = Unit